Enable kubebuilder multigroup structure, move api -> apis/cassandra, controllers -> controllers/cassandra
