+++ controllers/cassandra/suite_test.go
@@ -30,7 +30,7 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	cassandradatastaxcomv1beta1 "github.com/k8ssandra/cass-operator/api/v1beta1"
+	cassandradatastaxcomv1beta1 "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 	//+kubebuilder:scaffold:imports
 )
