+++ pkg/reconciliation/check_nodes.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/k8ssandra/cass-operator/pkg/utils"
 
-	api "github.com/k8ssandra/cass-operator/api/v1beta1"
+	api "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 )
 
 func (rc *ReconciliationContext) GetPodPVC(podNamespace string, podName string) (*corev1.PersistentVolumeClaim, error) {
