+++ pkg/reconciliation/construct_podtemplatespec_test.go
@@ -13,7 +13,7 @@ import (
 
 	"k8s.io/apimachinery/pkg/api/resource"
 
-	api "github.com/k8ssandra/cass-operator/api/v1beta1"
+	api "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 	"github.com/k8ssandra/cass-operator/pkg/images"
 	"github.com/k8ssandra/cass-operator/pkg/oplabels"
 	"github.com/stretchr/testify/assert"
