+++ pkg/reconciliation/constructor.go
@@ -6,7 +6,7 @@ package reconciliation
 // This file defines constructors for k8s objects
 
 import (
-	api "github.com/k8ssandra/cass-operator/api/v1beta1"
+	api "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 	"github.com/k8ssandra/cass-operator/pkg/oplabels"
 	"github.com/k8ssandra/cass-operator/pkg/utils"
