+++ pkg/reconciliation/reconcile_racks_helpers.go
@@ -5,7 +5,7 @@ package reconciliation
 import (
 	"fmt"
 
-	api "github.com/k8ssandra/cass-operator/api/v1beta1"
+	api "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 	"github.com/k8ssandra/cass-operator/pkg/httphelper"
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/core/v1"
