+++ tests/config_secret/config_secret_suite_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	api "github.com/k8ssandra/cass-operator/api/v1beta1"
+	api "github.com/k8ssandra/cass-operator/apis/cassandra/v1beta1"
 	"github.com/k8ssandra/cass-operator/tests/kustomize"
 	corev1 "k8s.io/api/core/v1"
